minor warnings/typefixes
git-svn-id: svn+ssh://frodo/cluster/svn/projects@5228 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ psc_fsutil_libs/psc_rpc/events.c
@@ -528,7 +528,8 @@ int pscrpc_ni_init(int type)
 			psc_fatalx("failed LNetNIInit() (%d)", rc);
 
 		rc = LNetEQAlloc(1024, pscrpc_master_callback, &pscrpc_eq_h);
-		psc_info("%#"PRIx64" pscrpc_eq_h cookie value", pscrpc_eq_h.cookie);
+		psc_info("%#"PRIx64" pscrpc_eq_h cookie value",
+		    (uint64_t)pscrpc_eq_h.cookie);
 	} else {
 		/* liblustre calls the master callback when it removes events from the
 		 * event queue.  The event queue has to be big enough not to drop
